# kbmk_backend-main/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'KBMK UNMUL API Documentation'
  description: 'Dokumentasi API untuk team Frontend'
  contact:
    email: admin@admin.com
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
servers:
  -
    url: /api
paths:
...
  /pengurus:
    get:
      tags:
        - Pengurus
      summary: 'Menampilkan daftar pengurus (dengan paginasi)'
      description: 'Endpoint publik untuk melihat data pengurus'
      operationId: 6819b06a19ae2b1a6a6b03047df521bb
      responses:
        200:
          description: Berhasil
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pengurus'
                  links:
                    type: object
                  meta:
                    properties:
                      current_page:
                        type: integer
                      last_page:
                        type: integer
                    type: object
                type: object
    post:
      tags:
        - Pengurus
      summary: 'Menambah data pengurus baru'
      description: "Memerlukan autentikasi dan role 'superadmin'. Mengupload foto."
      operationId: 968a70d4e1ea103510b99a7786f65fd0
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - nama
                - divisi
                - foto
                - deskripsi
              properties:
                nama:
                  type: string
                  maxLength: 50
                divisi:
                  type: string
                  maxLength: 50
                foto:
                  type: string
                  format: binary
                deskripsi:
                  type: string
              type: object
      responses:
        201:
          description: 'Berhasil dibuat'
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/Pengurus'
                type: object
        422:
          description: 'Error validasi'
        401:
          description: Unauthorized
      security:
        -
          bearerAuth: []
  '/pengurus/{id}':
    get:
      tags:
        - Pengurus
      summary: 'Menampilkan detail pengurus'
      description: 'Endpoint publik untuk melihat detail satu pengurus'
      operationId: 27a11dd2fee68d03ece0b0a0d431f565
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Berhasil
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/Pengurus'
                type: object
        404:
          description: 'Not found'
    put:
      tags:
        - Pengurus
      summary: 'Memperbarui data pengurus'
      description: "Memerlukan autentikasi dan role 'superadmin'. Dapat mengganti foto."
      operationId: 6b741e0346150b1b59fd643091749604
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                nama:
                  type: string
                  maxLength: 50
                divisi:
                  type: string
                  maxLength: 50
                foto:
                  type: string
                  format: binary
                deskripsi:
                  type: string
              type: object
      responses:
        200:
          description: 'Berhasil diperbarui'
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/Pengurus'
                type: object
        404:
          description: 'Not found'
        422:
          description: 'Error validasi'
        401:
          description: Unauthorized
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Pengurus
      summary: 'Menghapus data pengurus'
      description: "Memerlukan autentikasi dan role 'superadmin'"
      operationId: 369b620e24eddd2aa6077b574a5555f9
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Berhasil dihapus'
        404:
          description: 'Not found'
        401:
          description: Unauthorized
      security:
        -
          bearerAuth: []
  /profile-descs:
    get:
      tags:
        - 'Profile Descriptions'
      summary: 'Menampilkan semua deskripsi profil'
      operationId: 0826618d22f0f31b19cf36f31c9e9831
      responses:
        200:
          description: Berhasil
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProfileDesc'
                type: object
    post:
      tags:
        - 'Profile Descriptions'
      summary: 'Menambah deskripsi profil baru'
      description: "Memerlukan autentikasi dan role 'superadmin'"
      operationId: cf5735bc67e85435efb54d659b4c33ab
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileDesc'
      responses:
        201:
          description: 'Berhasil dibuat'
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ProfileDesc'
                type: object
        422:
          description: 'Error validasi'
        401:
          description: Unauthorized
      security:
        -
          bearerAuth: []
  '/profile-descs/{id}':
    get:
      tags:
        - 'Profile Descriptions'
      summary: 'Menampilkan detail deskripsi profil'
      operationId: f241d2a528350d8c650c82a31375f9d8
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Berhasil
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ProfileDesc'
                type: object
        404:
          description: 'Not found'
    put:
      tags:
        - 'Profile Descriptions'
      summary: 'Memperbarui deskripsi profil'
      description: "Memerlukan autentikasi dan role 'superadmin'"
      operationId: a6fd4e376899f51fc9dede4d71c97c7e
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                judul:
                  type: string
                  maxLength: 50
                jenis:
                  type: string
                  enum:
                    - about
                    - visi
                    - misi
                deskripsi:
                  type: string
              type: object
      responses:
        200:
          description: 'Berhasil diperbarui'
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ProfileDesc'
                type: object
        404:
          description: 'Not found'
        422:
          description: 'Error validasi'
        401:
          description: Unauthorized
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - 'Profile Descriptions'
      summary: 'Menghapus deskripsi profil'
      description: "Memerlukan autentikasi dan role 'superadmin'"
      operationId: 0b30fae05a4db71cfa8f6ea108ddb779
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Berhasil dihapus'
        404:
          description: 'Not found'
        401:
          description: Unauthorized
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    Pengurus:
      required:
        - nama
        - divisi
        - deskripsi
      properties:
        id:
          type: integer
          example: 1
        nama:
          type: string
          maxLength: 50
          example: 'Nanda Arianto'
        divisi:
          type: string
          maxLength: 50
          example: Humas
        foto:
          type: string
          example: pengurus/nama_file.jpg
        foto_url:
          type: string
          example: 'http://localhost:8000/storage/pengurus/nama_file.jpg'
        deskripsi:
          type: string
          example: 'Bertanggung jawab atas hubungan masyarakat.'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
    ProfileDesc:
      required:
        - judul
        - jenis
        - deskripsi
      properties:
        id:
          type: integer
          example: 1
        judul:
          type: string
          maxLength: 50
          example: 'Tentang Kami'
        jenis:
          type: string
          enum:
            - about
            - visi
            - misi
          example: about
        deskripsi:
          type: string
          example: 'Kami adalah organisasi yang...'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      bearerFormat: JWT
      scheme: bearer
...